+++ public/test/helpers/TestProvider.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { I18nProvider } from '../../app/core/localisation';
+import { I18nProvider } from '../../app/core/internationalization';
 
 const TestProvider: React.FC = ({ children }) => {
   return <I18nProvider>{children}</I18nProvider>;
